changing page layout
